boot: Switch to 32-bit protected mode and enter the kernel code.
Follow the standard procedure to shift to 32-bit protected mode so
that we can shift to C for writing the kernel code.

Enter the kernel code. The entry point of the kernel will be
kernel_main() function.

Note:-
Standard procedure requires to load the GDT, write the gdt in
gdt.asm file for now, later shift to C for writing GDT.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>